requirements.txt
js2py==0.68
requests==2.23.0
lxml==4.5.0
darkdetect==0.1.1
colorama==0.4.3
pycryptodome==3.9.8
pandas==1.0.5